% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{copepod_fetch}
\alias{copepod_fetch}
\title{Fecth COPEPD data from NMFS}
\usage{
copepod_fetch(
  name = "us-04201",
  dest = get_data_path(),
  decompress = TRUE,
  cleanup = decompress
)
}
\arguments{
\item{name}{character, the name of the data set ala "us-04201"}

\item{dest}{character, the path to save the data to}

\item{decompress}{logical, if TRUE unpack the downloaded zip file}

\item{cleanup}{logical, if TRUE remove the downloaded zip file. Ignored unless
\code{decompress} is \code{TRUE}}
}
\value{
the result of \code{\link[utils](download.file)}
}
\description{
Fecth COPEPD data from NMFS
}
