% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{copepod_shortform_names}
\alias{copepod_shortform_names}
\title{Retrieve a named vector of column names for copepod shortform}
\usage{
copepod_shortform_names(strip = TRUE)
}
\arguments{
\item{strip}{logical, if TRUE strip the names off}
}
\value{
named vector of column names
}
\description{
Retrieve a named vector of column names for copepod shortform
}
