% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{copepod_shortform_types}
\alias{copepod_shortform_types}
\title{Retrieve a named vector of column types for copepod shortform}
\usage{
copepod_shortform_types(compress = TRUE)
}
\arguments{
\item{compress}{logical, if TRUE transform into the \code{\link[readr]{read_delim}}
compact string codes}
}
\value{
named vector of column names or a character coding of such
}
\description{
Retrieve a named vector of column types for copepod shortform
}
