% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copepoda-package.R
\docType{package}
\name{copepoda}
\alias{copepoda}
\alias{copepod}
\title{copepoda - manage local data from NMFS plankton database}
\description{
Functions for reading [COPEPOD](https://www.st.nmfs.noaa.gov/plankton/about/about-copepod.html)
}
