% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{list_data}
\alias{list_data}
\title{Retrieve the path to one or more copepod datasets}
\usage{
list_data(
  name = "us-04201",
  form = c("short", "full")[1],
  root = get_data_path()
)
}
\description{
Retrieve the path to one or more copepod datasets
}
