% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_copepod}
\alias{read_copepod}
\title{Read a copepod short form - trim to bare necessities}
\usage{
read_copepod(
  filename = list_data(name = "us-05101"),
  simplify = TRUE,
  select_vars = c("date", "lon", "lat", "zupper", "zlower", "value_per_vol",
    "value_per_area", "lifestage"),
  form = c("tibble", "sf")[1]
)
}
\arguments{
\item{filename}{one or more filenames.  If multiples, the
individual tables are row bound into one.}

\item{simplify}{logical if TRUE trim down to a simple dataset}

\item{select_vars}{character, the columns to select if simplyfing.  Ignored
unless \code{simplify = TRUE}.}

\item{form}{character either 'tibble' or 'sf'}
}
\value{
tibble or sf Points object
}
\description{
Read a copepod short form - trim to bare necessities
}
