% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_copepod_shortform}
\alias{read_copepod_shortform}
\title{Read a short-form data file}
\usage{
read_copepod_shortform(
  filename,
  skip = 17,
  col_names = copepod_shortform_names(),
  col_types = copepod_shortform_types()
)
}
\arguments{
\item{filename}{the name of the file}

\item{skip}{the number of header lines to skip}

\item{col_names}{a vector of columns names including a trailing 'dummy' which is dropped}
}
\value{
tibble
}
\description{
Read a short-form data file
}
